Rename handleCallbackError file.
